requirements.txt
-r backend/requirements.txt
ruff# AI Integration
openai
python-dotenv
requests
cryptography
